Move express route to regex logic to separate file
